Initial commit to replace template values with the project-specific counterparts.
