+++ .github/dependabot.yml
@@ -9,12 +9,12 @@ version: 2
 # Ensure you have created a personal access token in GitHub and configured it with repo, write:packages and read:packages access.
 # The token must also be authorized with (MoJ) SSO. See link below for guidance:
 # https://docs.github.com/en/authentication/keeping-your-account-and-data-secure/managing-your-personal-access-tokens#creating-a-personal-access-token-classic
-#registries:
-#  spring-boot-common-github-packages:
-#    type: maven-repository
-#    url: https://maven.pkg.github.com/ministryofjustice/laa-ccms-spring-boot-common
-#    username: <your-username-here>
-#    password: ${{ secrets.REPO_TOKEN }}
+registries:
+  spring-boot-common-github-packages:
+    type: maven-repository
+    url: https://maven.pkg.github.com/ministryofjustice/laa-ccms-spring-boot-common
+    username: ${{ github.actor }}
+    password: ${{ secrets.GITHUB_TOKEN }}
 
 updates:
   - package-ecosystem: "gradle"
@@ -25,10 +25,10 @@ updates:
     groups:
       gradle-updates:
         exclude-patterns:
-          - "uk.gov.laa.springboot.microservice.*"
+          - "uk.gov.laa.datastewardship.payments.*"
       internal-packages:
         patterns:
-          - "uk.gov.laa.springboot.microservice*"
+          - "uk.gov.laa.datastewardship.payments*"
   - package-ecosystem: "github-actions"
     directory: "/"
     schedule:
