+++ .github/workflows/build-test-pr.yml
@@ -44,7 +44,7 @@ jobs:
         uses: actions/upload-artifact@v4
         with:
           name: checkstyle-report
-          path: spring-boot-microservice-service/build/reports/checkstyle
+          path: laa-data-stewardship-payments-service/build/reports/checkstyle
           retention-days: 14
 
       - name: Upload test report
@@ -52,7 +52,7 @@ jobs:
         uses: actions/upload-artifact@v4
         with:
           name: test-report
-          path: spring-boot-microservice-service/build/reports/tests
+          path: laa-data-stewardship-payments-service/build/reports/tests
           retention-days: 14
 
       - name: Upload jacoco coverage report
@@ -60,7 +60,7 @@ jobs:
         uses: actions/upload-artifact@v4
         with:
           name: jacoco-coverage-report
-          path: spring-boot-microservice-service/build/reports/jacoco
+          path: laa-data-stewardship-payments-service/build/reports/jacoco
           retention-days: 14
 
   vulnerability-scan-app:
@@ -104,10 +104,10 @@ jobs:
         run: ./gradlew build
 
       - name: Build a Docker image
-        run: docker build -t spring-boot-microservice:scan .
+        run: docker build -t laa-data-stewardship-payments:scan .
 
       - name: Run Snyk to check Docker image for vulnerabilities
         uses: snyk/actions/docker@master
         with:
-          image: spring-boot-microservice:scan
+          image: laa-data-stewardship-payments:scan
           args: --file=Dockerfile --severity-threshold=high
