+++ .gitignore
@@ -6,7 +6,7 @@ build/
 !**/src/test/**/build/
 .DS_Store
 
-spring-boot-microservice-api/generated
+laa-data-stewardship-payments-api/generated
 
 ### IntelliJ IDEA ###
 .idea
