+++ laa-data-stewardship-payments-service/src/test/java/uk/gov/justice/laa/datastewardship/payments/DataStewardshipPaymentsApplicationTests.java
@@ -1,0 +1,13 @@
+package uk.gov.justice.laa.datastewardship.payments;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class DataStewardshipPaymentsApplicationTests {
+
+  @Test
+  void contextLoads() {
+    // empty due to only testing context load
+  }
+}
