--- spring-boot-microservice-service/src/main/java/uk/gov/justice/laa/springboot/microservice/exception/GlobalExceptionHandler.java
@@ -1,41 +1,0 @@
-package uk.gov.justice.laa.springboot.microservice.exception;
-
-import static org.springframework.http.HttpStatus.NOT_FOUND;
-
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
-
-/**
- * The global exception handler for all exceptions.
- */
-@RestControllerAdvice
-@Slf4j
-public class GlobalExceptionHandler extends ResponseEntityExceptionHandler {
-
-  /**
-   * The handler for ItemNotFoundException.
-   *
-   * @param exception the exception
-   * @return the response status with error message
-   */
-  @ExceptionHandler(ItemNotFoundException.class)
-  public ResponseEntity<String> handleItemNotFound(ItemNotFoundException exception) {
-    return ResponseEntity.status(NOT_FOUND).body(exception.getMessage());
-  }
-
-  /**
-   * The handler for Exception.
-   *
-   * @param exception the exception
-   * @return the response status with error message
-   */
-  @ExceptionHandler(Exception.class)
-  public ResponseEntity<String> handleGenericException(Exception exception) {
-    String logMessage = "An unexpected application error has occurred.";
-    log.error(logMessage, exception);
-    return ResponseEntity.internalServerError().body(logMessage);
-  }
-}
