--- spring-boot-microservice-service/src/main/resources/application.yml
@@ -1,41 +1,0 @@
-spring:
-  application:
-    name: LAA Spring Boot Microservice
-
-  # example database
-  datasource:
-    url: jdbc:h2:mem:itemsDb
-    driver-class-name: org.h2.Driver
-    username: sa
-    password: password
-  h2:
-    console:
-      enabled: true
-
-  #do not auto create db schema, use .sql scripts
-  jpa:
-    hibernate:
-      ddl-auto: none
-
-# used by actuator info endpoint
-info:
-  app:
-    name: LAA Spring Boot Microservice application
-    description: An example Spring Boot Microservice application that provides a REST API.
-    version: 1.0.0
-
-management:
-  # specify actuator endpoints to enable
-  endpoints:
-    web:
-      exposure:
-        include: health,info
-
-  # show detailed health status
-  endpoint:
-    health:
-      show-details: always
-
-  info:
-    env:
-      enabled: true
