--- spring-boot-microservice-service/src/test/java/uk/gov/justice/laa/springboot/microservice/SpringBootMicroserviceApplicationTests.java
@@ -1,13 +1,0 @@
-package uk.gov.justice.laa.springboot.microservice;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class SpringBootMicroserviceApplicationTests {
-
-  @Test
-  void contextLoads() {
-    // empty due to only testing context load
-  }
-}
