--- spring-boot-microservice-service/src/test/java/uk/gov/justice/laa/springboot/microservice/exception/GlobalExceptionHandlerTest.java
@@ -1,33 +1,0 @@
-package uk.gov.justice.laa.springboot.microservice.exception;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.springframework.http.HttpStatus.INTERNAL_SERVER_ERROR;
-import static org.springframework.http.HttpStatus.NOT_FOUND;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.http.ResponseEntity;
-
-class GlobalExceptionHandlerTest {
-
-  GlobalExceptionHandler globalExceptionHandler = new GlobalExceptionHandler();
-
-  @Test
-  void handleItemNotFound_returnsNotFoundStatusAndErrorMessage() throws Exception {
-    ResponseEntity<String> result = globalExceptionHandler.handleItemNotFound(new ItemNotFoundException("Item not found"));
-
-    assertThat(result).isNotNull();
-    assertThat(result.getStatusCode()).isEqualTo(NOT_FOUND);
-    assertThat(result.getBody()).isNotNull();
-    assertThat(result.getBody()).isEqualTo("Item not found");
-  }
-
-  @Test
-  void handleGenericException_returnsInternalServerErrorStatusAndErrorMessage() throws Exception {
-    ResponseEntity<String> result = globalExceptionHandler.handleGenericException(new RuntimeException("Something went wrong"));
-
-    assertThat(result).isNotNull();
-    assertThat(result.getStatusCode()).isEqualTo(INTERNAL_SERVER_ERROR);
-    assertThat(result.getBody()).isNotNull();
-    assertThat(result.getBody()).isEqualTo("An unexpected application error has occurred.");
-  }
-}
